requirements.txt
streamlit
tensorflow
streamlit-drawable-canvas
pillow
opencv-python
numpy